+++ src/compilation.rs
@@ -319,7 +319,7 @@ mod test {
     zpr_address = "fd5a:5052:90de::1"
     interfaces = [ "in1" ]
     in1.netaddr = "127.0.0.1:5000"
-    provider = [["device.zpr.adapter.cn", "fee"]]
+    provider = [["endpoint.zpr.adapter.cn", "fee"]]
 
     [visa_service]
     dock_node = "n0"
@@ -338,8 +338,8 @@ mod test {
     #[test]
     fn simple_compile() {
         let zpl = r#"
-        define Webby as service with device.zpr.adapter.cn
-        allow zpr.adapter.cn: devices to access Webby
+        define Webby as service with endpoint.zpr.adapter.cn
+        allow zpr.adapter.cn: endpoints to access Webby
         "#;
 
         let tempdir = TempDir::new("simple_compile");
@@ -367,7 +367,7 @@ mod test {
     fn define_requires_with() {
         let zpl = r#"
         define Webby as service
-        allow zpr.adapter.cn: devices to access Webby
+        allow zpr.adapter.cn: endpoints to access Webby
         "#;
 
         let tempdir = TempDir::new("define_requires_with");
@@ -401,7 +401,7 @@ mod test {
         zpr_address = "fd5a:5052:90de::1"
         interfaces = [ "in1" ]
         in1.netaddr = "127.0.0.1:5000"
-        provider = [["device.zpr.adapter.cn", "fee"]]
+        provider = [["endpoint.zpr.adapter.cn", "fee"]]
 
         [visa_service]
         dock_node = "n0"
@@ -415,12 +415,12 @@ mod test {
 
         [services.Webby]
         protocol = "http"
-        provider = [["device.zpr.adapter.cn", ""]]
+        provider = [["endpoint.zpr.adapter.cn", ""]]
         "#;
 
         let zpl = r#"
         define Webby as service
-        allow zpr.adapter.cn: devices to access Webby
+        allow zpr.adapter.cn: endpoints to access Webby
         "#;
 
         let tempdir = TempDir::new("define_ok_without_with");
@@ -446,8 +446,8 @@ mod test {
     #[test]
     fn cannot_use_cn_as_tag() {
         let zpl = r#"
-        define Webby as service with device.zpr.adapter.cn
-        allow zpr.adapter.cn devices to access Webby
+        define Webby as service with endpoint.zpr.adapter.cn
+        allow zpr.adapter.cn endpoints to access Webby
         "#;
 
         let tempdir = TempDir::new("cannot_use_cn_as_tag");
@@ -476,7 +476,7 @@ mod test {
     fn test_svc_attrs_must_be_defined() {
         let zpl = r#"
         define Webby as service with unknown_attr
-        allow cn: devices to access services
+        allow cn: endpoints to access services
         "#;
 
         let tempdir = TempDir::new("test_svc_attrs_must_be_defined");
@@ -504,8 +504,8 @@ mod test {
     #[test]
     fn test_allow_attrs_must_be_defined() {
         let zpl = r#"
-        define Webby as service with device.zpr.adapter.cn
-        allow unknown_attr: devices to access services
+        define Webby as service with endpoint.zpr.adapter.cn
+        allow unknown_attr: endpoints to access services
         "#;
 
         let tempdir = TempDir::new("test_allow_attrs_must_be_defined");